refactor: refactor registration to use a single HTML page that can be configured with API data
