+++ config/webpack.config.js
@@ -12,7 +12,6 @@ const TerserPlugin = require('terser-webpack-plugin');
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 const OptimizeCSSAssetsPlugin = require('optimize-css-assets-webpack-plugin');
 const safePostCssParser = require('postcss-safe-parser');
-const ManifestPlugin = require('webpack-manifest-plugin');
 const InterpolateHtmlPlugin = require('react-dev-utils/InterpolateHtmlPlugin');
 const WorkboxWebpackPlugin = require('workbox-webpack-plugin');
 const WatchMissingNodeModulesPlugin = require('react-dev-utils/WatchMissingNodeModulesPlugin');
@@ -554,249 +553,14 @@ module.exports = function (webpackEnv) {
 				filename: 'login.html',
 			}),
 			new HtmlWebpackPlugin({
-				title: 'Registrierung Suchtberatung',
+				title: 'Registrierung',
 				templateParameters: {
 					type: 'registration',
-					consultingType: 0,
-					resortName: 'addiction',
 				},
 				inject: true,
 				template: 'src/pages/app.html',
 				chunks: ['registration'],
-				filename: 'registration.suchtberatung.html',
-			}),
-			new HtmlWebpackPlugin({
-				title:
-					'Registrierung Beratung für Suizidgefährdete junge Menschen [U25]',
-				templateParameters: {
-					type: 'registration',
-					consultingType: 1,
-					resortName: 'u25',
-				},
-				template: 'src/pages/app.html',
-				chunks: ['registration'],
-				filename: 'registration.u25.html',
-			}),
-			new HtmlWebpackPlugin({
-				title:
-					'Registrierung gemeinsamstatteinsam - Onlineberatung für junge Menschen in besonderen Zeiten',
-				templateParameters: {
-					type: 'registration',
-					consultingType: 1,
-					resortName: 'gemeinsamstatteinsam',
-				},
-				template: 'src/pages/app.html',
-				chunks: ['registration'],
-				filename: 'registration.gemeinsamstatteinsam.html',
-			}),
-			new HtmlWebpackPlugin({
-				title: 'Registrierung Schwangerschaftsberatung',
-				templateParameters: {
-					type: 'registration',
-					consultingType: 2,
-					resortName: 'pregnancy',
-				},
-				template: 'src/pages/app.html',
-				chunks: ['registration'],
-				filename: 'registration.schwangerschaftsberatung.html',
-			}),
-			new HtmlWebpackPlugin({
-				title: 'Registrierung Beratung für Eltern und Familien',
-				templateParameters: {
-					type: 'registration',
-					consultingType: 3,
-					resortName: 'parenting',
-				},
-				template: 'src/pages/app.html',
-				chunks: ['registration'],
-				filename: 'registration.eltern-familie.html',
-			}),
-			new HtmlWebpackPlugin({
-				title: 'Registrierung Kurberatung für Mütter und Väter',
-				templateParameters: {
-					type: 'registration',
-					consultingType: 4,
-					resortName: 'cure',
-				},
-				template: 'src/pages/app.html',
-				chunks: ['registration'],
-				filename: 'registration.kurberatung.html',
-			}),
-			new HtmlWebpackPlugin({
-				title: 'Registrierung Beratung zum Thema Schulden',
-				templateParameters: {
-					type: 'registration',
-					consultingType: 5,
-					resortName: 'debt',
-				},
-				template: 'src/pages/app.html',
-				chunks: ['registration'],
-				filename: 'registration.schuldnerberatung.html',
-			}),
-			new HtmlWebpackPlugin({
-				title: 'Registrierung Allgemeine Sozialberatung',
-				templateParameters: {
-					type: 'registration',
-					consultingType: 6,
-					resortName: 'social',
-				},
-				template: 'src/pages/app.html',
-				chunks: ['registration'],
-				filename: 'registration.allgemeine-soziale-beratung.html',
-			}),
-			new HtmlWebpackPlugin({
-				title: 'Registrierung Leben im Alter',
-				templateParameters: {
-					type: 'registration',
-					consultingType: 7,
-					resortName: 'seniority',
-				},
-				template: 'src/pages/app.html',
-				chunks: ['registration'],
-				filename: 'registration.leben-im-alter.html',
-			}),
-			new HtmlWebpackPlugin({
-				title: 'Registrierung Leben mit Behinderung und psychischer Erkrankung',
-				templateParameters: {
-					type: 'registration',
-					consultingType: 8,
-					resortName: 'disability',
-				},
-				template: 'src/pages/app.html',
-				chunks: ['registration'],
-				filename: 'registration.behinderung-und-psychische-erkrankung.html',
-			}),
-			new HtmlWebpackPlugin({
-				title: 'Registrierung Mein PlanB - Beratung zwischen Schule und Beruf',
-				templateParameters: {
-					type: 'registration',
-					consultingType: 9,
-					resortName: 'planB',
-				},
-				template: 'src/pages/app.html',
-				chunks: ['registration'],
-				filename: 'registration.mein-planb.html',
-			}),
-			new HtmlWebpackPlugin({
-				title: 'Registrierung Rechtliche Betreuung und Vorsorge',
-				templateParameters: {
-					type: 'registration',
-					consultingType: 10,
-					resortName: 'law',
-				},
-				template: 'src/pages/app.html',
-				chunks: ['registration'],
-				filename: 'registration.rechtliche-betreuung.html',
-			}),
-			new HtmlWebpackPlugin({
-				title: 'Registrierung Beratung für Angehörige von Straffälligen',
-				templateParameters: {
-					type: 'registration',
-					consultingType: 11,
-					resortName: 'offender',
-				},
-				template: 'src/pages/app.html',
-				chunks: ['registration'],
-				filename: 'registration.straffaelligkeit.html',
-			}),
-			new HtmlWebpackPlugin({
-				title: 'Registrierung HIV/AIDS Beratung',
-				templateParameters: {
-					type: 'registration',
-					consultingType: 12,
-					resortName: 'aids',
-				},
-				template: 'src/pages/app.html',
-				chunks: ['registration'],
-				filename: 'registration.hiv-aids-beratung.html',
-			}),
-			new HtmlWebpackPlugin({
-				title: 'Registrierung Kinder- und Jugend-Rehabilitation',
-				templateParameters: {
-					type: 'registration',
-					consultingType: 13,
-					resortName: 'rehabilitation',
-				},
-				template: 'src/pages/app.html',
-				chunks: ['registration'],
-				filename: 'registration.kinder-reha.html',
-			}),
-			new HtmlWebpackPlugin({
-				title: 'Registrierung Beratung für Kinder und Jugendliche',
-				templateParameters: {
-					type: 'registration',
-					consultingType: 14,
-					resortName: 'children',
-				},
-				template: 'src/pages/app.html',
-				chunks: ['registration'],
-				filename: 'registration.kinder-jugendliche.html',
-			}),
-			new HtmlWebpackPlugin({
-				title: 'Registrierung Sucht-Selbsthilfe - Kreuzbund-Chat',
-				templateParameters: {
-					type: 'registration',
-					consultingType: 15,
-					resortName: 'kreuzbund',
-				},
-				template: 'src/pages/app.html',
-				chunks: ['registration'],
-				filename: 'registration.kb-sucht-selbsthilfe.html',
-			}),
-			new HtmlWebpackPlugin({
-				title: 'Registrierung Migrationsberatung',
-				templateParameters: {
-					type: 'registration',
-					consultingType: 16,
-					resortName: 'migration',
-				},
-				template: 'src/pages/app.html',
-				chunks: ['registration'],
-				filename: 'registration.migration.html',
-			}),
-			new HtmlWebpackPlugin({
-				title: 'Registrierung Aus-/Rück- & Weiterwanderung',
-				templateParameters: {
-					type: 'registration',
-					consultingType: 17,
-					resortName: 'emigration',
-				},
-				template: 'src/pages/app.html',
-				chunks: ['registration'],
-				filename: 'registration.rw-auswanderung.html',
-			}),
-			new HtmlWebpackPlugin({
-				title: 'Registrierung Hospiz-, Palliativ- und Trauerberatung',
-				templateParameters: {
-					type: 'registration',
-					consultingType: 18,
-					resortName: 'hospice',
-				},
-				template: 'src/pages/app.html',
-				chunks: ['registration'],
-				filename: 'registration.hospiz-palliativ.html',
-			}),
-			new HtmlWebpackPlugin({
-				title: 'Registrierung Regionale Angebote',
-				templateParameters: {
-					type: 'registration',
-					consultingType: 19,
-					resortName: 'regional',
-				},
-				template: 'src/pages/app.html',
-				chunks: ['registration'],
-				filename: 'registration.regionale-angebote.html',
-			}),
-			new HtmlWebpackPlugin({
-				title: 'Registrierung Jungen- und Männerberatung',
-				templateParameters: {
-					type: 'registration',
-					consultingType: 20,
-					resortName: 'men',
-				},
-				template: 'src/pages/app.html',
-				chunks: ['registration'],
-				filename: 'registration.jungen-und-maenner.html',
+				filename: 'registration.html',
 			}),
 			new HtmlWebpackPlugin({
 				title: 'Error Page 401',
