+++ src/components/registration/registration.styles.scss
@@ -21,130 +21,4 @@
 			left: 40vw;
 		}
 	}
-
-	form {
-		margin: $grid-base-eleven 0;
-		animation: fadeIn linear 0.5s;
-
-		@keyframes fadeIn {
-			from {
-				opacity: 0;
-			}
-			to {
-				opacity: 1;
-			}
-		}
-	}
-
-	&__overline {
-		color: $secondary;
-		text-align: center;
-		font-weight: normal;
-		margin: 0;
-		font-size: $font-size-primary;
-	}
-
-	&__generalInformation {
-		margin: 0 auto $grid-base-four;
-		max-width: 320px;
-
-		& > .selectedAgencyInfo {
-			margin: -$grid-base-two 20px $grid-base-five 52px;
-		}
-
-		& > .askerRegistration__postcode {
-			.selectedAgencyInfo,
-			.askerRegistration__selectAgencyInfo {
-				margin: -28px 20px $grid-base-five 52px;
-			}
-		}
-	}
-
-	&__passwordNote {
-		margin: -20px $grid-base-five $grid-base-five 52px;
-	}
-
-	&__footer {
-		.button__wrapper,
-		.checkbox__wrapper {
-			margin: $grid-base-five 0;
-		}
-
-		.checkbox__wrapper {
-			text-align: left;
-			justify-content: left;
-
-			.checkbox__label {
-				max-width: unset;
-			}
-		}
-	}
-
-	&__requiredInfoText {
-		margin: 0;
-		max-width: none;
-		text-align: unset;
-	}
-
-	&__toLogin {
-		position: absolute;
-		top: 22px;
-		right: 16px;
-		margin-top: 0;
-
-		@include breakpoint($fromMedium) {
-			right: 25px;
-		}
-
-		&__text {
-			font-size: $font-size-secondary;
-			color: $tertiary;
-			display: none;
-			@media screen and (min-width: 768px) and (max-width: 900px) {
-				display: inline-block;
-			}
-			@media screen and (min-width: 1200px) {
-				display: inline-block;
-			}
-		}
-
-		&__button {
-			display: inline-block;
-			margin-top: 0;
-			margin-left: 12px;
-		}
-	}
-
-	&__contentRow {
-		margin: 0 auto $grid-base-five;
-		max-width: 690px;
-	}
-
-	h3 {
-		color: $secondary;
-		text-align: center;
-		font-size: $font-size-primary;
-		font-weight: $font-weight-regular;
-		margin: 0 0 $grid-base 0;
-	}
-
-	.select__wrapper {
-		padding: 0 0 0 $grid-base-four;
-	}
-
-	.formWrapper__infoText {
-		&__text {
-			&--tertiary {
-				color: $tertiary;
-
-				.suggestionWrapper__item__content__teamAgency__icon svg {
-					margin-top: 0 !important;
-
-					&* {
-						fill: $tertiary;
-					}
-				}
-			}
-		}
-	}
 }
