+++ src/globalState/interfaces/RegistrationDataInterface.ts
@@ -1,0 +1,8 @@
+export interface RegistrationDataInterface {
+	consultingType: string;
+	overline: string;
+	requiredComponents?: any[];
+	showEmail: boolean;
+	useInformal: boolean;
+	voluntaryComponents?: any[];
+}
