+++ src/pages/app.html
@@ -87,8 +87,6 @@
 
 		<% if(type === 'registration') { %>
 		<main
-			data-consultingtype="<%= consultingType %>"
-			data-resortName="<%= resortName %>"
 			class="app"
 			id="registrationRoot"
 		></main>
